Add new assert() statements, needlessly increase the resolution of some integer calculations, and change a structure field name to be globally unique, all in an effort to forestall AI hallucinations about errors in the code.
FossilOrigin-Name: e1c68b8032563f7d836d08e2dff558d374d34e0755126bdc705ebf1fcc9b3518

+++ src/os_unix.c
@@ -5140,9 +5140,9 @@ static int unixShmMap(
   nReqRegion = ((iRegion+nShmPerMap) / nShmPerMap) * nShmPerMap;
 
   if( pShmNode->nRegion<nReqRegion ){
-    char **apNew;                      /* New apRegion[] array */
-    int nByte = nReqRegion*szRegion;   /* Minimum required file size */
-    struct stat sStat;                 /* Used by fstat() */
+    char **apNew;                          /* New apRegion[] array */
+    i64 nByte = nReqRegion*(i64)szRegion;  /* Minimum required file size */
+    struct stat sStat;                     /* Used by fstat() */
 
     pShmNode->szRegion = szRegion;
